# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /deleteToken:
    x-swagger-router-controller: login
    put:
      tags:
        - "set-token-null"
      description: set token null that was already logged in
      operationId: deleteAdminToken
      parameters:
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: empty
          description: empty body
          schema:
            type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /allAlliances:
    x-swagger-router-controller: alliance
    get:
      description: get the alliances list
      operationId: getAllAlliances
      tags:
        - "get alliances"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AlliancesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alliances:
    x-swagger-router-controller: alliance
    get:
      description: get the alliances list
      operationId: getAlliances
      tags:
        - "get alliances"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AlliancesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new alliance to the list
      operationId: postAlliance
      tags:
        - "post alliances"
      parameters:
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: alliance
          description: The alliance to create.
          schema:
            $ref: "#/definitions/Alliance"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alliance/{id}:
    x-swagger-router-controller: alliance
    get:
      description: get an alliance
      operationId: getAlliance
      parameters: 
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema: 
           $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update an alliance
      operationId: updateAlliance
      # define the parameters
      tags:
        - "update Alliance"
      parameters:
        - name: id
          description: alliance id
          type: number
          in: path
          required: true
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: alliance
          description: The alliance to update.
          schema:
            $ref: "#/definitions/Alliance"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateAllianceState/{id}:
    x-swagger-router-controller: alliance
    put:
      tags:
        - "update Alliance"
      description: Update state of alliance
      operationId: updateStateAlliance
      parameters:
        - name: id 
          description: alliance id 
          type: string
          in: path
          required: true
        - in: header
          name: token
          description: alliance token
          type: string
          required: true
        - in: body 
          name: state
          schema:
            $ref: "#/definitions/State"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /complaints:
    x-swagger-router-controller: complaint
    get:
      description: get the complaints list
      operationId: getComplaints
      tags:
        - "get complaints"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ComplaintListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new complaint to the list
      operationId: postComplaint
      tags:
        - "post complaint"
      parameters:
        - in: body
          name: complaint
          description: The complaint to create.
          schema:
            $ref: "#/definitions/Complaint"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /complaints/{id}:
    x-swagger-router-controller: complaint
    get:
      description: get a complaint
      operationId: getComplaint
      parameters: 
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema: 
           $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a complaint
      operationId: updateComplaint
      # define the parameters
      tags:
        - "update Complaint"
      parameters:
        - name: id
          description: complaint id
          type: number
          in: path
          required: true
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: complaint
          description: The complaint to update.
          schema:
            $ref: "#/definitions/User"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    delete:
      description: delete a complaint
      # define the parameters
      parameters:
        - name: id
          description: complaint id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralDeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GeneralDeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  State:
    type: object
    properties:
      state:
        type: boolean
...
  GeneralObjectResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
    required:
      - id
...
  Alliance:
    type: object
    properties:
      name:
        type: string
        description: Name of the alliance
      description:
        type: string
        description: Description of the alliance
      offer:
        type: string
        description: Offer terms and conditions
      website:
        type: string
        description: Website of the alliance
      email:
        type: string
        description: Email of the alliance
      phone:
        type: string
        description: Phone of the alliance
      state:
        type: boolean
        description: indicates if the alliance is 'noted' or no
      finish_date:
        type: string
        description: Closing date of the offer
      image:
        type: string
        format: byte
    required:
      - name
      - description
      - offer
      - finish_date
      - email
      - state
      - image
  AlliancesListResponse:
    required:
      - alliances
    properties:
      alliances:
        type: array
        items:
          $ref: "#/definitions/GeneralObjectResponse"
  Complaint:
    type: object
    properties:
      first_name:
        type: string
        description: First name of the complainant
      last_name:
        type: string
        description: Last name of the complainant
      document_type:
        type: string
        description: Type of document
      document_number:
        type: string
        description: Document number
      email:
        type: string
        description: complainant's mail
      phone:
        type: string
        description: complainant's phone
      event_day:
        type: string
        description: Date on which the event occurred
      event_place:
        type: string
        description: Place where the event occurred
      description:
        type: string
        description: Description of the situation of vulnerability
    required:
      - first_name
      - last_name
      - document_type
      - document_number
      - email
      - phone
      - event_day
      - event_place
      - description
  ComplaintListResponse:
    required:
      - complaints
    properties:
      complaints:
        type: array
        items:
          $ref: "#/definitions/GeneralObjectResponse"
  User:
    type: object
    properties:
      document_type:
        type: string
        description: Type of user document
      document_number:
        type: string
        description: User's document number
      first_name:
        type: string
        description: User's first name
      last_name:
        type: string
        description: User's last name
      address:
        type: string
        description: User's Address 
      email:
        type: string
        description: User's email
      phone:
        type: string
        description: User's phone
      sex_birth:
        type: string
        description: User's birth sex
      sexual_orientation:
        type: string
        description: Sexual orientation of the user
      gender:
        type: string
        description: Gender of the user
      birth_day:
        type: string
        description: User's birth day      
      education:
        type: string
        description: User's education
    required:
      - document_type
      - document_number
      - first_name
      - last_name
      - address
      - email
      - phone
      - sex_birth
      - sexual_orientation
      - gender
      - birth_day
      - education
...